Update README.md
Robert has moved into a new house. He has a number N of boxes from his previous 
home. Each box bi has a volume vi. He must fit those boxes in 2 rooms of the house and is not 
allowed to put any box outside of those two rooms yet. The volume of the room1 is V1 and the 
volume of the room2 is V2. The total volume of the boxes (v1+…+vn) is greater than V1+V2. So, he 
is not going to be able to fit all the boxes, but he wants to maximize the numbers of boxes in the 
2 rooms (or minimize the boxes left outside of the rooms). He is an expert in robotics and got 2 
robotic agents that can do the moving for them. The agents can move around the 2 rooms and 
the house, they can detect the boxes and their size with their sensors and can get them and place 
them in the rooms. 
The successor function embedded in the software that controls the agents has been defined to 
pick up boxes at random and place them in any of the rooms also at random until there is no 
more space in the rooms. You need to define a new successor function by using an algorithm we 
have covered in class and hand the implementation to Robert, so he can insert it in the software 
that control the robots.